requirements.txt
mysql-connector-python==8.0.31
protobuf==3.20.1
python-dotenv==0.21.0
camel_converter==3.0.2
python-dateutil==2.8.2
python-dotenv==1.0.0
